requirements.txt
anthropic>=0.8.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
jinja2>=3.1.0
matplotlib>=3.8.0
numpy>=1.24.0
openai>=1.3.0
pandas>=2.1.0
pathspec>=0.11.0
pillow>=10.1.0
plotly>=5.18.0
pyperclip>=1.8.0
pydantic>=2.5.0
requests>=2.31.0
scikit-learn>=1.3.0
scipy>=1.11.0
tiktoken>=0.5.0
